[Feature-16992][dolphinscheduler-task-seatunnel] add jvm option in front-end form
