+++ dolphinscheduler-ui/src/views/projects/task/components/node/tasks/use-sea-tunnel.ts
@@ -94,7 +94,9 @@ export function useSeaTunnel({
     parallelism: 1,
     sourceCustomParams: [],
     targetCustomParams: [],
-    jobMode: 'BATCH'
+    jobMode: 'BATCH',
+    xms: 1,
+    xmx: 1
   } as INodeData)
 
   return {
